Modifications to .cpp files to generate RcppExports.cpp
I had to add a series of arma:: inside some of the .cpp scripts because I was trying to "install and restart" R using R studio. The way it was before, the program would not compile and I got several error messages mainly asking to add arma:: to some variables (vec, mat and uvec). It also returned some error messages regarding some comments inside the definition of a function parameters. I have specifically added coments inside the scripts when that was the came. My comment started as //FFN  comment.